+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Adding: attr_reader for `:path, :html_options, :width, :height`
 - Adding: `keywords.json` and `codepoints.json` methods to the octicon object. Making them available via the api
+- Deprecating: the use of `:size` for "large" or a number, instead use `:height`
 
 # 1.1.0